rearange file
