+++ file_parser/range_check_ambient.c
@@ -6,7 +6,7 @@
 /*   By: ferafano <ferafano@student.42antananarivo  +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/12/07 13:18:47 by ferafano          #+#    #+#             */
-/*   Updated: 2024/12/07 13:43:05 by ferafano         ###   ########.fr       */
+/*   Updated: 2024/12/07 14:00:54 by ferafano         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,7 +15,6 @@
 int valid_range_ambient(char **str2)
 {
 	double value;
-	float amb_light_ratio;
 
 	value = 0.0;
 	if (is_valid_float(str2[1]))
